refactor-components * Move LoginModal to components directory
