+++ src/components/LoginModal/EmailLogin/EmailLogin.tsx
@@ -7,11 +7,11 @@ import { BackArrowIcon, ClosedLockIcon, EmailIcon } from '@entur/icons'
 import { PrimaryButton } from '@entur/button'
 import { Heading3, Link } from '@entur/typography'
 import { SmallExpandableAlertBox } from '@entur/alert'
-import { auth } from '../../../../UserProvider'
-import { useFormFields } from '../../../../utils'
-import sikkerhetBom from '../../../../assets/images/sikkerhet_bom.png'
-import retinaSikkerhetBom from '../../../../assets/images/sikkerhet_bom@2x.png'
-import { CloseButton } from '../../../CloseButton/CloseButton'
+import { auth } from '../../../UserProvider'
+import { useFormFields } from '../../../utils'
+import sikkerhetBom from '../../../assets/images/sikkerhet_bom.png'
+import retinaSikkerhetBom from '../../../assets/images/sikkerhet_bom@2x.png'
+import { CloseButton } from '../../CloseButton/CloseButton'
 import { ModalType } from '../login-modal-types'
 
 export interface UserLogin {
