+++ src/components/LoginModal/EmailSent/EmailSent.tsx
@@ -3,9 +3,9 @@ import type { User } from 'firebase/auth'
 import { GridContainer, GridItem } from '@entur/grid'
 import { PrimaryButton } from '@entur/button'
 import { Heading3, Paragraph } from '@entur/typography'
-import { CloseButton } from '../../../CloseButton/CloseButton'
-import Check from '../../../../assets/images/check.png'
-import retinaCheck from '../../../../assets/images/check@2x.png'
+import { CloseButton } from '../../CloseButton/CloseButton'
+import Check from '../../../assets/images/check.png'
+import retinaCheck from '../../../assets/images/check@2x.png'
 import { ModalType } from '../login-modal-types'
 
 interface EmailSentProps {
