+++ src/components/LoginModal/LoginModal.scss
@@ -1,4 +1,4 @@
-@import '../../../assets/styles/import.scss';
+@import '../../assets/styles/import.scss';
 
 .eds-modal__overlay {
     background: var(--tavla-modal-overlay-color);
