+++ src/components/LoginModal/LoginModal.tsx
@@ -2,8 +2,8 @@ import React, { useCallback, useEffect, useState } from 'react'
 import type { User } from 'firebase/auth'
 import { useToast } from '@entur/alert'
 import { Modal } from '@entur/modal'
-import { useUser } from '../../../UserProvider'
-import { usePrevious } from '../../../utils'
+import { useUser } from '../../UserProvider'
+import { usePrevious } from '../../utils'
 import { EmailLogin } from './EmailLogin/EmailLogin'
 import { LoginOptions } from './LoginOptions/LoginOptions'
 import { Signup } from './Signup/Signup'
