+++ src/components/LoginModal/ResetPassword/ResetPassword.tsx
@@ -6,11 +6,11 @@ import { GridContainer, GridItem } from '@entur/grid'
 import { BackArrowIcon, EmailIcon } from '@entur/icons'
 import { PrimaryButton } from '@entur/button'
 import { Heading3, Paragraph } from '@entur/typography'
-import { auth } from '../../../../UserProvider'
-import { useFormFields } from '../../../../utils'
-import sikkerhetBom from '../../../../assets/images/sikkerhet_bom.png'
-import retinaSikkerhetBom from '../../../../assets/images/sikkerhet_bom@2x.png'
-import { CloseButton } from '../../../CloseButton/CloseButton'
+import { auth } from '../../../UserProvider'
+import { useFormFields } from '../../../utils'
+import sikkerhetBom from '../../../assets/images/sikkerhet_bom.png'
+import retinaSikkerhetBom from '../../../assets/images/sikkerhet_bom@2x.png'
+import { CloseButton } from '../../CloseButton/CloseButton'
 import { ModalType } from '../login-modal-types'
 
 export interface UserResetPassword {
