+++ src/components/Modals/MineTavlerModal/MineTavlerModal.tsx
@@ -6,8 +6,8 @@ import { GridContainer, GridItem } from '@entur/grid'
 import { PrimaryButton, SecondaryButton } from '@entur/button'
 import { useUser } from '../../../UserProvider'
 import { useSettings } from '../../../settings/SettingsProvider'
-import { LoginModal } from '../LoginModal/LoginModal'
-import { LoginCase } from '../LoginModal/login-modal-types'
+import { LoginModal } from '../../LoginModal/LoginModal'
+import { LoginCase } from '../../LoginModal/login-modal-types'
 import { CloseButton } from '../../CloseButton/CloseButton'
 import sikkerhetBom from '../../../assets/images/sikkerhet_bom.png'
 import retinaSikkerhetBom from '../../../assets/images/sikkerhet_bom@2x.png'
