+++ src/containers/Admin/LogoTab/LogoTab.tsx
@@ -11,8 +11,8 @@ import {
 import { GridContainer, GridItem } from '@entur/grid'
 import { useUser } from '../../../UserProvider'
 import { getDocumentId } from '../../../utils'
-import { LoginModal } from '../../../components/Modals/LoginModal/LoginModal'
-import { LoginCase } from '../../../components/Modals/LoginModal/login-modal-types'
+import { LoginModal } from '../../../components/LoginModal/LoginModal'
+import { LoginCase } from '../../../components/LoginModal/login-modal-types'
 import { LogoUpload } from './LogoUpload/LogoUpload'
 import { SizePicker } from './SizePicker/SizePicker'
 import { Description } from './Description/Description'
