+++ src/containers/Admin/NameTab/NameTab.tsx
@@ -4,8 +4,8 @@ import { Heading2, Paragraph } from '@entur/typography'
 import { GridContainer, GridItem } from '@entur/grid'
 import { useUser } from '../../../UserProvider'
 import { getDocumentId } from '../../../utils'
-import { LoginModal } from '../../../components/Modals/LoginModal/LoginModal'
-import { LoginCase } from '../../../components/Modals/LoginModal/login-modal-types'
+import { LoginModal } from '../../../components/LoginModal/LoginModal'
+import { LoginCase } from '../../../components/LoginModal/login-modal-types'
 import './NameTab.scss'
 import { CustomURL } from './CustomURL/CustomURL'
