+++ src/containers/DashboardWrapper/BottomMenu/BottomMenu.tsx
@@ -15,8 +15,8 @@ import { useToast } from '@entur/alert'
 import { useSettings } from '../../../settings/SettingsProvider'
 import { auth, useUser } from '../../../UserProvider'
 import { LockModal } from '../../../components/Modals/LockModal/LockModal'
-import { LoginModal } from '../../../components/Modals/LoginModal/LoginModal'
-import { LoginCase } from '../../../components/Modals/LoginModal/login-modal-types'
+import { LoginModal } from '../../../components/LoginModal/LoginModal'
+import { LoginCase } from '../../../components/LoginModal/login-modal-types'
 import { MineTavlerModal } from '../../../components/Modals/MineTavlerModal/MineTavlerModal'
 import { MenuButton } from './MenuButton/MenuButton'
 import './BottomMenu.scss'
