+++ src/containers/Error/ErrorPages.tsx
@@ -2,8 +2,8 @@ import React, { useState } from 'react'
 import { useNavigate } from 'react-router-dom'
 import { signOut } from 'firebase/auth'
 import { useToast } from '@entur/alert'
-import { LoginModal } from '../../components/Modals/LoginModal/LoginModal'
-import { LoginCase } from '../../components/Modals/LoginModal/login-modal-types'
+import { LoginModal } from '../../components/LoginModal/LoginModal'
+import { LoginCase } from '../../components/LoginModal/login-modal-types'
 import { auth, useUser } from '../../UserProvider'
 import { getDocumentId } from '../../utils'
 import sikkerhetBomLight from '../../assets/images/sikkerhet_bom_light@2x.png'
